design(스토리북에 글로벌 스타일, 테마 적용): #13
스토리북에 초기 css 세팅한 글로벌 스타일과 테마를 적용시키기 위해 Decorator에 ThemeProvider, GlobalStyle을 설정하였습니다.
